Refactored: Moved the helper functions in indexers.indexer_config to indexers.utils.
