+++ medusa/helpers/externals.py
@@ -6,8 +6,9 @@ import logging
 
 from medusa import app, db
 from medusa.indexers.indexer_api import indexerApi
-from medusa.indexers.indexer_config import indexerConfig, mappings
+from medusa.indexers.indexer_config import indexerConfig
 from medusa.indexers.indexer_exceptions import IndexerException, IndexerShowAllreadyInLibrary, IndexerUnavailable
+from medusa.indexers.utils import mappings
 from medusa.logger.adapters.style import BraceAdapter
 
 from requests.exceptions import RequestException
