+++ medusa/metadata/generic.py
@@ -9,7 +9,6 @@ from medusa import app, exception_handler, helpers
 from medusa.helper.common import replace_extension
 from medusa.helper.exceptions import ex
 from medusa.helper.metadata import get_image
-from medusa.indexers.indexer_api import indexerApi
 from medusa.indexers.indexer_config import INDEXER_TMDB, INDEXER_TVDBV2, INDEXER_TVMAZE
 from medusa.indexers.indexer_exceptions import (IndexerEpisodeNotFound, IndexerException,
                                                 IndexerSeasonNotFound, IndexerShowNotFound)
@@ -733,7 +732,7 @@ class GenericMetadata(object):
         if image_type not in (u'fanart', u'poster', u'banner', u'thumbnail', u'poster_thumb', u'banner_thumb'):
             log.error(
                 u'Invalid {image}, unable to find it in the {indexer}',
-                {u'image': image_type, u'indexer': indexerApi(show_obj.indexer).name}
+                {u'image': image_type, u'indexer': show_obj.indexer_name}
             )
             return None
 
@@ -855,14 +854,14 @@ class GenericMetadata(object):
         except IndexerShowNotFound:
             log.warning(
                 u'Unable to find {indexer} show {id}, skipping it',
-                {u'indexer': indexerApi(series_obj.indexer).name, u'id': series_id}
+                {u'indexer': series_obj.indexer_name, u'id': series_id}
             )
             return False
 
         except (IndexerException, RequestException):
             log.warning(
                 u'{indexer} is down, cannot use its data to add this show',
-                {u'indexer': indexerApi(series_obj.indexer).name}
+                {u'indexer': series_obj.indexer_name}
             )
             return False
 
@@ -870,7 +869,7 @@ class GenericMetadata(object):
         if not (getattr(my_show, u'seriesname', None) and getattr(my_show, u'id', None)):
             log.warning(
                 u'Incomplete info for {indexer} show {id}, skipping it',
-                {u'indexer': indexerApi(series_obj.indexer).name, u'id': series_id}
+                {u'indexer': series_obj.indexer_name, u'id': series_id}
             )
             return False
