+++ medusa/server/web/manage/handler.py
@@ -32,7 +32,7 @@ from medusa.helper.exceptions import (
     CantUpdateShowException,
 )
 from medusa.helpers import is_media_file
-from medusa.indexers.indexer_config import indexer_id_to_name, indexer_name_to_id
+from medusa.indexers.utils import indexer_id_to_name, indexer_name_to_id
 from medusa.network_timezones import app_timezone
 from medusa.post_processor import PostProcessor
 from medusa.server.web.core import PageTemplate, WebRoot
